// meu-projeto/src/models/FormData.ts
export interface FormData {
    codigo_campo: string;
    codigo_cooperado: string;
    rensasem: string;
    nome_produtor: string;
    safra: string;
    cooperado: string;
    propriedade: string;
    inscricao_estadual: string;
    municipio: string;
    uf: string;
    responsavel: string;
    assistente: string;
    codigo_sap: string;
    cultivar: string;
    cpfCnpj: string;
  }  